+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased changes
+- Added support for Initialising smart contract
+
 
 ## 3.1.0
 
@@ -37,7 +39,6 @@
 - Added support for query `UnBanNode`
 - Added support for query `StartBaker`
 - Added support for query `StopBaker`
-- Added support for Initialising smart contract
 
 ## 3.0.0
 - Stronger typing for various places in the API.